Refactor: re-order methods and divide tests into files.
